refactor: implement logout functionality in AccountPage and removed logout code from GroupsList
